rework of all cmds
refactored all commands into their own js files for a cleaner directory approach
